src/consumer/requirements.txt
kafka-python==2.0.2
google-cloud-bigquery==3.9.0
requests==2.31.0